restructure account app layout
